api/requirements.txt
aiofiles
aiosmtplib
aiosqlite
alembic
anyio
asgiref
asyncpg
attrs
autopep8
bcrypt
blinker
certifi
cffi
cfgv
charset-normalizer
coverage
cryptography
databases
distlib
dnspython
ecdsa
email-validator
fastapi
fastapi-mail
fastapi-offline
fastapi-pagination
filelock
greenlet
gunicorn
h11
identify
idna
iniconfig
Jinja2
Mako
MarkupSafe
mccabe
mypy-extensions
nodeenv
packaging
passlib
pathspec
Pillow
platformdirs
pluggy
psycopg2-binary
py
pyasn1
pycodestyle
pycparser
pydantic
pyflakes
PyJWT
pyparsing
pytest
pytest-cov
python-dotenv
python-jose
python-multipart
PyYAML
requests
rsa
six
sniffio
SQLAlchemy
SQLAlchemy-Utils
sqlalchemy2-stubs
sqlmodel
starlette
toml
tomli
typing_extensions
ujson
urllib3
uvicorn
virtualenv
websockets
